requirements.txt
Django==4.2.7
yt-dlp>=2024.10.7
moviepy==1.0.3
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-generativeai==0.3.1
Pillow>=10.3.0
numpy>=2.0.0
scipy>=1.14.0
librosa>=0.10.2
numba>=0.60.0
soundfile>=0.12.1
opencv-python>=4.8.0
